organized some stuff
