+++ app/src/main/AndroidManifest.xml
@@ -11,10 +11,10 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".fragments.activity.ActivityTwo">
+        <activity android:name=".activity.ActivityTwo">
 
         </activity>
-        <activity android:name=".fragments.activity.MainActivity">
+        <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
