+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:id="@+id/main_fragment_holder"
-tools:context=".fragments.activity.MainActivity">
+tools:context=".activity.MainActivity">
